+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 # Changelog
 
+## v1.18.6
+
+[compare changes](https://github.com/Zotlo/zotlo-checkout/compare/v1.18.5...v1.18.6)
+
+### 🏡 Chore
+
+- Prevent merchant validation if applepay already cancelled ([55fe63a](https://github.com/Zotlo/zotlo-checkout/commit/55fe63a))
+
+### ❤️ Contributors
+
+- Sinan Mutlu ([@SinanMtl](https://github.com/SinanMtl))
+
+## v1.18.5
+
+[compare changes](https://github.com/Zotlo/zotlo-checkout/compare/v1.18.4...v1.18.5)
+
+### 🏡 Chore
+
+- Handle service response on apple merchant validation error ([2979e4a](https://github.com/Zotlo/zotlo-checkout/commit/2979e4a))
+
+### ❤️ Contributors
+
+- Sinan Mutlu ([@SinanMtl](https://github.com/SinanMtl))
+
 ## v1.18.4
 
 [compare changes](https://github.com/Zotlo/zotlo-checkout/compare/v1.18.3...v1.18.5)